refactor: change fastAPI structure to pratice suggested by github.com/zhanymkanov/fastapi-best-practices
