+++ backend/src/news/service.py
@@ -4,18 +4,9 @@ from urllib.parse import quote
 import requests
 from bs4 import BeautifulSoup
 from openai import OpenAI
-from pydantic import BaseModel
 from sqlalchemy.orm import Session
 
-from .database import NewsArticleRepository
-
-
-class PromptRequest(BaseModel):
-    prompt: str
-
-
-class NewsSummaryRequestSchema(BaseModel):
-    content: str
+from ..database import NewsArticleRepository
 
 
 class NewsProcessor:
@@ -27,15 +18,6 @@ class NewsProcessor:
     def _fetch_news_from_api(
         self, search_term: str, is_initial: bool = False
     ) -> list[dict]:
-        """
-        Fetch news metadata from UDN's search API.
-
-        :param search_term: The term to search for.
-        :param is_initial: If True, fetches multiple pages (1..9);
-          else only page 1.
-        :return: A list of raw news objects from the API
-          (contains title, titleLink, etc.).
-        """
         all_news_data: list[dict] = []
         pages_to_fetch = range(1, 10) if is_initial else range(1, 2)
 
@@ -61,12 +43,6 @@ class NewsProcessor:
         return all_news_data
 
     def _check_relevance(self, title: str) -> bool:
-        """
-        Ask the model to judge if the title is about
-        「民生用品的價格變化」.
-
-        Returns True only if model answers 'high'.
-        """
         messages = [
             {
                 "role": "system",
@@ -93,14 +69,6 @@ class NewsProcessor:
             return False
 
     def _scrape_article_details(self, url: str) -> dict | None:
-        """
-        Given a UDN article URL, scrape:
-        - title
-        - time
-        - content paragraphs (list[str])
-
-        Returns dict on success, None on failure.
-        """
         try:
             response = requests.get(url, timeout=10)
             response.raise_for_status()
@@ -112,8 +80,6 @@ class NewsProcessor:
                 "section", class_="article-content__editor"
             )
 
-            # If structure changed and we can't find required nodes,
-            # bail out gracefully
             if not title_el or not time_el or not content_section:
                 print(f"Missing expected elements while scraping {url}")
                 return None
@@ -136,14 +102,6 @@ class NewsProcessor:
             return None
 
     def _summarize_article(self, content_blocks: list[str]) -> dict | None:
-        """
-        Ask the model to summarize:
-        - 影響
-        - 原因
-
-        Expects pure valid JSON with keys "影響" and "原因".
-        Returns {"summary": "...", "reason": "..."} or None.
-        """
         messages = [
             {
                 "role": "system",
@@ -178,40 +136,28 @@ class NewsProcessor:
             return None
 
     def process_news(self, is_initial: bool = False):
-        """
-        Fetches, filters, summarizes, and stores news articles
-        about price changes in 民生用品.
-
-        :param is_initial: If True, crawl multiple pages (broader fetch).
-        """
         news_items = self._fetch_news_from_api("價格", is_initial)
 
         for item in news_items:
-            # item should contain "title" and "titleLink" from UDN API
             title = item.get("title")
             url = item.get("titleLink")
             if not title or not url:
                 continue
 
-            # Step 1: filter by relevance
             if not self._check_relevance(title):
                 continue
 
-            # Step 2: scrape full article
             detailed_news = self._scrape_article_details(url)
             if not detailed_news:
                 continue
 
-            # Step 3: summarize with LLM
             summary_data = self._summarize_article(detailed_news["content"])
             if not summary_data:
                 continue
 
-            # Attach summary metadata required by DB insert
             detailed_news["summary"] = summary_data["summary"]
             detailed_news["reason"] = summary_data["reason"]
 
-            # Step 4: persist
             self.news_repo.add_article(detailed_news)
 
             print(f"Added article: {detailed_news['title']}")
