+++ backend/src/user/router.py
@@ -1,0 +1,47 @@
+from fastapi import APIRouter, Depends, HTTPException, status
+from fastapi.security import OAuth2PasswordRequestForm
+from sqlalchemy.orm import Session
+
+from ..database import User, get_db
+from .schemas import Token, UserAuthSchema
+from .service import auth_service
+
+router = APIRouter(prefix="/api/v1/users", tags=["users"])
+
+
+@router.post("/login", response_model=Token)
+async def login_for_access_token(
+    form_data: OAuth2PasswordRequestForm = Depends(),
+    db: Session = Depends(get_db),
+):
+    user = auth_service.authenticate_user(
+        db, form_data.username, form_data.password
+    )
+    access_token = auth_service.create_access_token(
+        data={"sub": user.username}
+    )
+    return {"access_token": access_token, "token_type": "bearer"}
+
+
+@router.post("/register", status_code=status.HTTP_201_CREATED)
+def create_user(user: UserAuthSchema, db: Session = Depends(get_db)):
+    db_user = db.query(User).filter(User.username == user.username).first()
+    if db_user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Username already registered",
+        )
+    hashed_password = auth_service.get_password_hash(user.password)
+    db_user = User(username=user.username, hashed_password=hashed_password)
+    db.add(db_user)
+    db.commit()
+    db.refresh(db_user)
+    return {
+        "username": db_user.username,
+        "message": "User created successfully",
+    }
+
+
+@router.get("/me")
+def read_users_me(current_user: User = Depends(auth_service.get_current_user)):
+    return {"username": current_user.username}
