+++ backend/tests/news/test_news_endpoint.py
